Add support for licensing generator for Rust directories
* Rust directories have a different structure where the BUILD.gn and
the README.chromium are in a different directory than the actual source
code. Soong requires that the Android.bp and METADATA live in the same
directory and Android.bp can't break the package boundary so we will
have to generate the METADATA + licensing files in the source directory.

* Found a bunch of violations that don't have a license file, filed a
bug for those and delegated to the rust team in Chromium to solve b/369075726.

* Added a test to ensure the correctness over the longterm.

Bug: 369300145
Change-Id: Iadc674379f1ce50624dc2dcd27b4d8c90eff235e