Ensure symlinks are relative and not absolute for LICENSE file
This bug was found when we added support for rust where the LICENSE file
path is an absolute path so the symlinks were pointing at absolute path
which is incorrect. This CL fixes it by ensuring that everything is
relative path.

Added a single line to the rust tests to ensure that the tests cover
this scenario.

Bug: 369300145
Change-Id: I8cef9dce40d7aabe981051a3eed06742ebe95442