Add support for rust_library and rust_proc_macro
The following new fields have been added:

* crate_name: Holds the name of the crate and is not arch-specific.
* crate_root: Holds the root file of the crate and is not arch-specific.
* rustlibs: Holds the rust dependencies of a module and is arch-specific.
* proc_macros: Holds the proc macros dependencies and is arch-specific.
* edition: Holds the edition of rustc compiler and is arch-specific.
* features: Holds the features to be used when compiling and is arch
 specific.
* cfgs: Holds the configs to be used when compiling and is arch specific.
* flags: Holds the rustflags to be passed to the compiler and is arch
specific.

It's important to note that rust is not really an architecture-specific
language, as all of the generated modules seems to have the same values
across different variants. We could move `edition` to be non-arch specific
in the future.

Removed `out` from the keys to check for arch-variant as Soong allows
genrules to specify arch-variant `out` as seen here:

https://android-review.git.corp.google.com/c/platform/build/soong/+/3002539

Bug: 346568402
Test: None
Change-Id: Id207ab92970b9251cc943407d1408405ce1e45cd