Disable failing licensing test due to b/372449684
The test is failing because every file in the `data` gets added to
the binary and we are exceeding the maximum length possible for a
command line length. It's fine to skip the test as the license tools
have been upstreamed to Chromium and will start running there instead
to catch bugs early.

Change-Id: I5b566580cb801ddf303995fbc91d64037e76b929