Build Cronet against the System SDK
This is essentially a revert of https://r.android.com/3189650. The
rationale is the current setup is causing problems with container
dependency checks; see https://r.android.com/3392583, b/382299968.

The reason we can simply revert this without breaking anything is
because the current setup was originally intended to support private
Connectivity API use in Cronet, but this capability is not actually
used yet. We'll have to revisit this again when we actually want to
start using private Connectivity APIs.

Test: m
Bug: 381421702
Bug: 382299968
Bug: 311792075
Change-Id: I34ce2e02820af7783b504676b937a768ec1af784