Preload Cronet's shared library into Zygote
After several A/B experiments runs, we found out that there's no
regression on boot-time to preload Cronet into the zygote (See
b/380349437#comment19).

The ApiVersion is equal to the ImplVersion, hence it's safe to load
the shared library by fetching its version through the ApiVersion class.

Test: Tested locally by flashing an image with preloaded shared library
and managed to send a GET request to www.google.com, verified that the
loading time has decreased.
Bug: 380349437

Change-Id: Idb36cd69fbefbe32dd8d207b1af11931a254f100