Add preupload hook to external/cronet
This should standardize the code formatting that is used across the
entire repository for directories that are not automatically imported
from Chromium (eg: android/, test_runner/).

The default for java-formatting is the google_java_format --aosp.
The default for CPP formatting is clang-format.
The default for kotlin is ktfmt.

Added automatic bpfmt on repo upload.

Removed the requirement to add 'Test:'.

Change-Id: Ia71f2069be0464698231f0d74075316419cd05fc