Stop jarjaring Cronet MTS tests
Cronet MTS Tests are our only source of coverage and confidence that
HttpEngine is working as intended, we were initially JarJaring those
tests to match what we ship in the platform so we could test the
platform code via this channel. However, this is no longer necessary
because we have HttpEngineWrapper which tests the platform code via
CronetEngine codepath without JarJar.

JarJar has also caused headache when importing as we have to make some
classes public instead of package-private in order to access them after
we JarJar. There's also the complexity of the blurry line of "is the
code being executed in the Platform or inside the APK?".

Finally, the JarJaring setup will completely break once we try to roll
https://crrev.com/c/6226162 as it assumes there's a single shared
library whose name is `libcronet_tests` which will not be bundled
in the platform while we'll be asking platform to do that.

Bug: 394583271
Change-Id: I159a00b5d7f6eacb85afbe5198a33dae28093d4b