Stop overriding flags in CTS tests
Overriding flags seems to be explicitly disabled as per the
[documentation](https://g3doc.corp.google.com/company/teams/android-core-experiments/trunk-stable/class-and-method-level-flag-override.md?cl=head).

xTS tests are supposed to be flag-agnostic which was not the case and
this has caused b/402303876 to occur, the alternative here is to use flag filtering. However, we think that the tests are no longer necessary at this point due to the fact that they run the default behavior which is what the current tests already do; We removed the tests.

Tested: m cts && run cts -m CtsNetHttpTestCases -t android.net.http.cts.HttpEngineTest
Bug: 402303876
Change-Id: I122dfe774f35dfa4d796b20f6c488bd102bddc8d